adding django script
